refactor: alias::JsonValue
