+++ src/eth/rpc/rpc_server.rs
@@ -54,7 +54,7 @@ use crate::eth::Consensus;
 use crate::ext::not;
 use crate::ext::to_json_string;
 use crate::ext::to_json_value;
-use crate::ext::JsonValue;
+use crate::alias::JsonValue;
 use crate::infra::build_info;
 use crate::infra::metrics;
 use crate::infra::tracing::SpanExt;
