+++ src/eth/storage/external_rpc_storage.rs
@@ -17,7 +17,7 @@ use crate::eth::primitives::Wei;
 use crate::eth::storage::PostgresExternalRpcStorage;
 use crate::eth::storage::PostgresExternalRpcStorageConfig;
 use crate::ext::parse_duration;
-use crate::ext::JsonValue;
+use crate::alias::JsonValue;
 
 #[async_trait]
 pub trait ExternalRpcStorage: Send + Sync {
